Model_API/requirements.txt
Flask==3.0.3
google-cloud-firestore==2.16.0
google-cloud-storage==2.17.0
keras==2.13.1
pillow==10.3.0
numpy==1.24.3
gunicorn==22.0.0
tensorflow==2.13.0